+++ api/controllers/admin/facebook/FeedController.js
@@ -1,0 +1,9 @@
+module.exports = {
+  index: async (req, res) => {
+    res.ok({
+      view: true,
+      serverSidePaging: true,
+      layout: 'admin/default/index'
+    });
+  }
+}