+++ api/models/Message.js
@@ -11,7 +11,8 @@ module.exports = {
       'orderConfirm',
       'paymentConfirm',
       'deliveryConfirm',
-      'orderSync'
+      'orderSync',
+      'forgotPassword'
     ),
     from: Sequelize.STRING,
     to: Sequelize.STRING,