+++ migration/changelog.md
@@ -4,6 +4,22 @@
 
 ### User 欄位變動
 
-* issue link
+
 
 ### Recipe 欄位變動
+
+
+## v2
+
+### Feed 欄位變動
+
+* 新增 createdTime 紀錄 feed 在 facebook 的建立時間
+* 新增 publish 紀錄 feed 是否發布
+
+### Message 欄位變動
+
+*  type 新增 forgotPassword enum
+
+### User 欄位變動
+s
+* 新增 resetPasswordToken 紀錄忘記密碼的 token